Move dbQueriesHelpers.js into lib directory
